requirements.txt
feedparser
deep-translator
tqdm
requests
beautifulsoup4
yfinance
python-dotenv